规范化HomePage代码结构
