+++ app/src/main/java/com/yenaly/han1meviewer/ui/component/PageContent.kt
@@ -5,6 +5,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import com.yenaly.han1meviewer.logic.state.PageState
 import com.yenaly.han1meviewer.ui.component.content.EmptyContent
 import com.yenaly.han1meviewer.ui.component.content.ErrorContent
 import com.yenaly.han1meviewer.ui.component.content.LoadingContent
@@ -37,8 +38,17 @@ fun PageContent(
 ) {
     when {
         isError && isEmpty -> error()
-        isLoading && isEmpty -> loading()
+        isLoading -> loading()
         isEmpty -> empty()
         else -> content()
     }
 }
+
+val <T> PageState<T>.isFirstPageLoading: Boolean
+    get() = this is PageState.Loading
+
+val <T> PageState<T>.isFirstPageError: Boolean
+    get() = this is PageState.Error && cachedInfo == null
+
+val <T> PageState<T>.isFirstPageEmpty: Boolean
+    get() = this is PageState.Empty
