+++ app/src/main/java/com/yenaly/han1meviewer/ui/screen/home/homepage/HomePageViewModel.kt
@@ -1,0 +1,212 @@
+package com.yenaly.han1meviewer.ui.screen.home.homepage
+
+import android.util.Log
+import androidx.annotation.StringRes
+import androidx.core.content.edit
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.google.firebase.database.FirebaseDatabase
+import com.yenaly.han1meviewer.FIREBASE_REALTIME_DATABASE
+import com.yenaly.han1meviewer.Preferences
+import com.yenaly.han1meviewer.R
+import com.yenaly.han1meviewer.SAVED_USER_ID
+import com.yenaly.han1meviewer.logic.DatabaseRepo
+import com.yenaly.han1meviewer.logic.NetworkRepo
+import com.yenaly.han1meviewer.logic.entity.HKeyframeEntity
+import com.yenaly.han1meviewer.logic.entity.WatchHistoryEntity
+import com.yenaly.han1meviewer.logic.exception.LoginStateExpiredException
+import com.yenaly.han1meviewer.logic.model.Announcement
+import com.yenaly.han1meviewer.logic.state.PageState
+import com.yenaly.han1meviewer.logic.state.WebsiteState
+import com.yenaly.han1meviewer.logout
+import com.yenaly.han1meviewer.ui.viewmodel.AppViewModel
+import com.yenaly.yenaly_libs.utils.getSpValue
+import com.yenaly.yenaly_libs.utils.putSpValue
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.async
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.catch
+import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.suspendCancellableCoroutine
+import kotlin.coroutines.resume
+
+class HomePageViewModel: ViewModel() {
+    data class SessionExpiredMessage(
+        val message: String?,
+        @param:StringRes val fallbackResId: Int,
+    )
+
+    private val _homePageFlow = MutableStateFlow<PageState<HomeData>>(PageState.Loading)
+    val homePageFlow = _homePageFlow.asStateFlow()
+
+    private val database = FirebaseDatabase.getInstance(FIREBASE_REALTIME_DATABASE)
+
+    private val _sessionExpiredMessage = MutableSharedFlow<SessionExpiredMessage>()
+    val sessionExpiredMessage = _sessionExpiredMessage
+
+    private val _announcements = MutableLiveData<List<Announcement>>()
+
+    init {
+        viewModelScope.launch {
+            // 初始化默认已下载分组，防止[FOREIGN KEY constraint failed]
+            DatabaseRepo.HanimeDownload.insertDefaultGroup()
+        }
+    }
+
+    fun getHomePage(isRefresh: Boolean = false){
+        viewModelScope.launch {
+            val current = _homePageFlow.value
+            if (isRefresh && current is PageState.Success){
+                _homePageFlow.value = current.copy(isRefreshing = true)
+            } else if (!isRefresh && current !is PageState.Success){
+                _homePageFlow.value = PageState.Loading
+            }
+            val announcementsDeferred = async(Dispatchers.IO) { fetchAnnouncementsFromFirebase() }
+            NetworkRepo.getHomePage().collect { networkState ->
+                when (networkState){
+                    is WebsiteState.Error -> {
+                        if (networkState.throwable is LoginStateExpiredException) {
+                            logout()
+                            _sessionExpiredMessage.emit(
+                                SessionExpiredMessage(
+                                    message = networkState.throwable.message,
+                                    fallbackResId = R.string.login_state_expired,
+                                )
+                            )
+                        }
+                        val previousData = (_homePageFlow.value as? PageState.Success)?.info
+                        _homePageFlow.value = PageState.Error(networkState.throwable, cachedInfo = previousData)
+                    }
+                    is WebsiteState.Success -> {
+                        val currentAnnouncements = announcementsDeferred.await()
+                        AppViewModel.csrfToken = networkState.info.csrfToken
+                        networkState.info.userId.takeIf { it.isNotEmpty() }?.let { userId ->
+                            Preferences.preferenceSp.edit { putString(SAVED_USER_ID, userId) }
+                        }
+                        val homeData = HomeData(page = networkState.info, announcements = currentAnnouncements)
+                        _homePageFlow.value = PageState.Success(info = homeData, isRefreshing = false)
+                    }
+                    is WebsiteState.Loading -> { }
+                }
+            }
+        }
+    }
+    private suspend fun fetchAnnouncementsFromFirebase(): List<Announcement> =
+        suspendCancellableCoroutine { continuation ->
+            val lastDismissTime = getSpValue("last_dismiss_time", 0L, "setting_pref")
+            val shouldShowAnno = System.currentTimeMillis() - lastDismissTime > 24 * 60 * 60 * 1000L
+            if (!shouldShowAnno) {
+                continuation.resume(emptyList())
+                return@suspendCancellableCoroutine
+            }
+
+            database.getReference("announcements").get()
+                .addOnSuccessListener { snapshot ->
+                    val list = mutableListOf<Announcement>()
+                    if (snapshot.exists()) {
+                        for (announceSnap in snapshot.children) {
+                            val announcement = announceSnap.getValue(Announcement::class.java)
+                            if (announcement != null && announcement.isActive) {
+                                list.add(announcement)
+                            }
+                        }
+                        continuation.resume(list.sortedBy { it.priority })
+                    } else {
+                        continuation.resume(emptyList())
+                    }
+                }.addOnFailureListener { e ->
+                    Log.e("Announcement", "读取失败: ${e.message}")
+                    continuation.resume(emptyList()) // 失败也容错返回空列表
+                }
+        }
+
+    fun dismissAnnouncements(){
+        putSpValue("last_dismiss_time", System.currentTimeMillis(), "setting_pref")
+        val current = _homePageFlow.value
+        if (current is PageState.Success) {
+            _homePageFlow.value = current.copy(info = current.info.copy(announcements = emptyList()))
+        }
+    }
+
+    fun deleteWatchHistory(history: WatchHistoryEntity) {
+        viewModelScope.launch(Dispatchers.IO) {
+            DatabaseRepo.WatchHistory.delete(history)
+            Log.d("delete_watch_hty", "$history DONE!")
+        }
+    }
+
+    fun deleteAllWatchHistories() {
+        viewModelScope.launch(Dispatchers.IO) {
+            DatabaseRepo.WatchHistory.deleteAll()
+            Log.d("del_all_watch_hty", "DONE!")
+        }
+    }
+
+    fun loadAllWatchHistories() =
+        DatabaseRepo.WatchHistory.loadAll()
+            .catch { e -> e.printStackTrace() }
+            .flowOn(Dispatchers.IO)
+    private val _modifyHKeyframeFlow = MutableSharedFlow<Boolean>()
+    fun removeHKeyframe(videoCode: String, hKeyframe: HKeyframeEntity.Keyframe) {
+        viewModelScope.launch(Dispatchers.IO) {
+            DatabaseRepo.HKeyframe.removeKeyframe(videoCode, hKeyframe)
+            Log.d("HKeyframe", "removeHKeyframe:$hKeyframe DONE!")
+            _modifyHKeyframeFlow.emit(true)
+        }
+    }
+    fun modifyHKeyframe(
+        videoCode: String,
+        oldKeyframe: HKeyframeEntity.Keyframe, keyframe: HKeyframeEntity.Keyframe,
+    ) {
+        viewModelScope.launch {
+            DatabaseRepo.HKeyframe.modifyKeyframe(videoCode, oldKeyframe, keyframe)
+            Log.d("HKeyframe", "modifyHKeyframe:$keyframe DONE!")
+            _modifyHKeyframeFlow.emit(true)
+        }
+    }
+    fun deleteHKeyframes(entity: HKeyframeEntity) {
+        viewModelScope.launch(Dispatchers.IO) {
+            DatabaseRepo.HKeyframe.delete(entity)
+        }
+    }
+
+    fun updateHKeyframes(entity: HKeyframeEntity) {
+        viewModelScope.launch(Dispatchers.IO) {
+            DatabaseRepo.HKeyframe.update(entity)
+        }
+    }
+    fun loadAnnouncements(forceRefresh: Boolean = false) {
+        val lastDismissTime = getSpValue("last_dismiss_time", 0L, "setting_pref")
+        val shouldShowAnno = System.currentTimeMillis() - lastDismissTime > 24*60*60*1000L
+        if (!shouldShowAnno) {
+            _announcements.value = emptyList()
+            return
+        }
+        if (_announcements.value != null && !forceRefresh) return
+
+        val announcementsRef = database.getReference("announcements")
+
+        announcementsRef.get().addOnSuccessListener { snapshot ->
+            val list = mutableListOf<Announcement>()
+            if (snapshot.exists()) {
+                for (announceSnap in snapshot.children) {
+                    val announcement = announceSnap.getValue(Announcement::class.java)
+                    if (announcement != null && announcement.isActive) {
+                        list.add(announcement)
+                    }
+                }
+                _announcements.postValue(list.sortedBy { it.priority })
+            } else {
+                _announcements.postValue(emptyList())
+            }
+            Log.i("Announcement",list.toString())
+        }.addOnFailureListener { e ->
+            Log.e("Announcement", "读取失败: ${e.message}")
+            _announcements.postValue(emptyList())
+        }
+    }
+}
