--- app/src/main/java/com/yenaly/han1meviewer/ui/viewmodel/MainViewModel.kt
@@ -1,166 +1,0 @@
-package com.yenaly.han1meviewer.ui.viewmodel
-
-import android.app.Application
-import android.util.Log
-import androidx.core.content.edit
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.viewModelScope
-import com.google.firebase.database.FirebaseDatabase
-import com.yenaly.han1meviewer.FIREBASE_REALTIME_DATABASE
-import com.yenaly.han1meviewer.Preferences
-import com.yenaly.han1meviewer.R
-import com.yenaly.han1meviewer.SAVED_USER_ID
-import com.yenaly.han1meviewer.logout
-import com.yenaly.han1meviewer.logic.DatabaseRepo
-import com.yenaly.han1meviewer.logic.exception.LoginStateExpiredException
-import com.yenaly.han1meviewer.logic.NetworkRepo
-import com.yenaly.han1meviewer.logic.entity.HKeyframeEntity
-import com.yenaly.han1meviewer.logic.entity.WatchHistoryEntity
-import com.yenaly.han1meviewer.logic.model.Announcement
-import com.yenaly.han1meviewer.logic.model.HomePage
-import com.yenaly.han1meviewer.logic.state.WebsiteState
-import com.yenaly.han1meviewer.ui.viewmodel.AppViewModel.csrfToken
-import com.yenaly.yenaly_libs.base.YenalyViewModel
-import com.yenaly.yenaly_libs.utils.getSpValue
-import com.yenaly.yenaly_libs.utils.putSpValue
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.launch
-
-/**
- * @project Hanime1
- * @author Yenaly Liew
- * @time 2022/06/08 008 17:35
- */
-class MainViewModel(application: Application) : YenalyViewModel(application) {
-
-    private val _homePageFlow =
-        MutableStateFlow<WebsiteState<HomePage>>(WebsiteState.Loading)
-    val homePageFlow = _homePageFlow.asStateFlow()
-    val horizontalScrollPositions = mutableMapOf<String, Int>()
-    private val _announcements = MutableLiveData<List<Announcement>>()
-    val announcements: LiveData<List<Announcement>> = _announcements
-    private val database = FirebaseDatabase.getInstance(FIREBASE_REALTIME_DATABASE)
-    private val _sessionExpiredMessage = MutableSharedFlow<String>()
-    val sessionExpiredMessage = _sessionExpiredMessage
-
-    init {
-        viewModelScope.launch {
-            // 初始化默认已下载分组，防止[FOREIGN KEY constraint failed]
-            DatabaseRepo.HanimeDownload.insertDefaultGroup()
-        }
-    }
-    fun getHomePage() {
-        viewModelScope.launch {
-            NetworkRepo.getHomePage().collect { homePage ->
-                if (homePage is WebsiteState.Error && homePage.throwable is LoginStateExpiredException) {
-                    logout()
-                    _sessionExpiredMessage.emit(
-                        homePage.throwable.message ?: application.getString(R.string.login_state_expired)
-                    )
-                }
-                if (homePage is WebsiteState.Success) {
-                    csrfToken = homePage.info.csrfToken
-                    homePage.info.userId.takeIf { it.isNotEmpty() }?.let { userId ->
-                        Preferences.preferenceSp.edit {
-                            putString(SAVED_USER_ID, userId)
-                        }
-                    }
-                }
-                _homePageFlow.value = homePage
-            }
-        }
-    }
-
-    fun deleteWatchHistory(history: WatchHistoryEntity) {
-        viewModelScope.launch(Dispatchers.IO) {
-            DatabaseRepo.WatchHistory.delete(history)
-            Log.d("delete_watch_hty", "$history DONE!")
-        }
-    }
-
-    fun deleteAllWatchHistories() {
-        viewModelScope.launch(Dispatchers.IO) {
-            DatabaseRepo.WatchHistory.deleteAll()
-            Log.d("del_all_watch_hty", "DONE!")
-        }
-    }
-
-    fun loadAllWatchHistories() =
-        DatabaseRepo.WatchHistory.loadAll()
-            .catch { e -> e.printStackTrace() }
-            .flowOn(Dispatchers.IO)
-    private val _modifyHKeyframeFlow = MutableSharedFlow<Pair<Boolean, String>>()
-    fun removeHKeyframe(videoCode: String, hKeyframe: HKeyframeEntity.Keyframe) {
-        viewModelScope.launch(Dispatchers.IO) {
-            DatabaseRepo.HKeyframe.removeKeyframe(videoCode, hKeyframe)
-            Log.d("HKeyframe", "removeHKeyframe:$hKeyframe DONE!")
-            _modifyHKeyframeFlow.emit(true to application.getString(R.string.delete_success))
-        }
-    }
-    fun modifyHKeyframe(
-        videoCode: String,
-        oldKeyframe: HKeyframeEntity.Keyframe, keyframe: HKeyframeEntity.Keyframe,
-    ) {
-        viewModelScope.launch {
-            DatabaseRepo.HKeyframe.modifyKeyframe(videoCode, oldKeyframe, keyframe)
-            Log.d("HKeyframe", "modifyHKeyframe:$keyframe DONE!")
-            _modifyHKeyframeFlow.emit(true to application.getString(R.string.modify_success))
-        }
-    }
-    fun deleteHKeyframes(entity: HKeyframeEntity) {
-        viewModelScope.launch(Dispatchers.IO) {
-            DatabaseRepo.HKeyframe.delete(entity)
-        }
-    }
-
-    fun updateHKeyframes(entity: HKeyframeEntity) {
-        viewModelScope.launch(Dispatchers.IO) {
-            DatabaseRepo.HKeyframe.update(entity)
-        }
-    }
-    fun loadAnnouncements(forceRefresh: Boolean = false) {
-        val lastDismissTime = getSpValue("last_dismiss_time",0L,"setting_pref")
-        val shouldShowAnno = System.currentTimeMillis() - lastDismissTime > 24*60*60*1000L
-        if (!shouldShowAnno) {
-            _announcements.value = emptyList()
-            return
-        }
-        if (_announcements.value != null && !forceRefresh) return
-
-        val announcementsRef = database.getReference("announcements")
-
-        announcementsRef.get().addOnSuccessListener { snapshot ->
-            val list = mutableListOf<Announcement>()
-            if (snapshot.exists()) {
-                for (announceSnap in snapshot.children) {
-                    val announcement = announceSnap.getValue(Announcement::class.java)
-                    if (announcement != null && announcement.isActive) {
-                        list.add(announcement)
-                    }
-                }
-                _announcements.postValue(list.sortedBy { it.priority })
-            } else {
-                _announcements.postValue(emptyList())
-            }
-            Log.i("Announcement",list.toString())
-        }.addOnFailureListener { e ->
-            Log.e("Announcement", "读取失败: ${e.message}")
-            _announcements.postValue(emptyList())
-        }
-    }
-
-    fun dismissAnnouncements() {
-        putSpValue(
-            "last_dismiss_time",
-            System.currentTimeMillis(),
-            "setting_pref"
-        )
-        _announcements.value = emptyList()
-    }
-}
